Add back sleep, but 1s instead of 3s, after vpnkit startup.
Originally there was a 3s sleep after relay, vpnkit and tap startup (totally 9s),
but this was changed upstream to tight wait loops instead - but then only for relay
and tap, not for vpnkit. This results in some suspicious messages on startup, because
vpnkit-tap-vsockd tries to negotiate a connection to vpnkit before it is ready.
It is not an actual error though, the messages just indicate that one attempt failed,
and vpnkit-tap-vsockd will just try again later, and succeed once vpnkit is up and running.

Example of messages one failed attempt of vpnkit-tap-vsockd connecting to vpnkit:

2021/09/21 11:48:27 open //./pipe/wsl-vpnkit: The system cannot find the file specified.
EOF reading from socket: closing

Failed to read hello from client
Failed to negotiate vmnet connection